Remove test code from lib/auth/keystore (#56609)
The package required testing packages for two reasons

1) To expose helpers which populated servicecfg.KeystoreConfig
   based on environment variables set by the host running the tests.
2) To expose a helper for creating a manager that used software
   keys for tests

The first case has been eliminated by copying the helpers to all
the other packages that called it. The helpers are all self contained
and don't rely on any keystore package specific logic. The duplication
was chosen instead of dependency for simplicity.

keystore.NewSoftwareKeystoreForTests existed only to optionally
allow overriding how RSA keys were generated. The same means has
been moved into the keystore.Options allowing all tests to call
keystore.NewManager directly.

Updates #51023.
